[spark] Introduce source.split.target-size-with-column-pruning (#6837)

+++ paimon-spark/paimon-spark-common/src/main/scala/org/apache/paimon/spark/scan/PaimonStatistics.scala
@@ -68,7 +68,8 @@ case class PaimonStatistics(
     }
   }
 
-  lazy val readRowSizeRatio: Double = estimateRowSize(readRowType) / estimateRowSize(tableRowType)
+  lazy val readRowSizeRatio: Double =
+    estimateRowSize(readRowType).toDouble / estimateRowSize(tableRowType)
 
   private def estimateRowSize(rowType: RowType): Long = {
     rowType.getFields.asScala.map(estimateFieldSize).sum
